Add files via upload
In this document, I explain how I computed biological aging (BA) measures using the BioAge package developed by Kwon et al. (Kwon and Belsky 2021). The goals of this analysis were (a) to compute the BA with a set of working biomarkers to represent Colombian population; and (b) to compare different BA prediction measures to elucidate the most fitting and applicable for the SABE dataset.